refactor(connectors): dedup file-source helpers and remove dead code
Hoist the duplicated dirExists/writePerRefManifest/readPerRefManifest/
pruneOldRefDirs helpers (tarball + local file sources) into the shared
cache.ts, using the stricter 64-hex ref-dir predicate. Remove dead code:
the singular generateEmbedding (only mock stubs), the void Readable
no-op, and the unused PollResponse watcher fields the daemon rejects.
Simplify pathExists (drop the failing readFile-first probe and dynamic
import), the resolver future-scheme loop, and the single-case scoring
switch. Behavior-preserving; typecheck clean, tests green.
